+++ src/utils/PatchNote.ts
@@ -4,7 +4,7 @@ import { Locale, ValError } from "@valapi/lib";
 
 // interface
 
-export namespace PatchNote {
+export namespace ValPatchNote {
     export type Version = `${number}.${number}`;
 }
 
@@ -13,10 +13,11 @@ export namespace PatchNote {
 /**
  * Valorant Patch Note URL
  */
-export class PatchNote {
-    private json: string;
-    private patch_list: string;
-    private patch_note: string;
+export class ValPatchNote {
+    private readonly baseUrl: {
+        default: string;
+        data: string;
+    };
 
     /**
      *
@@ -25,36 +26,13 @@ export class PatchNote {
     private constructor(language: Locale.Identify = "en-US") {
         const lowerCase_language = language.toLowerCase();
 
-        this.json = `https://playvalorant.com/page-data/${lowerCase_language}/news/tags/patch-notes/page-data.json`;
-
-        this.patch_list = `https://playvalorant.com/${lowerCase_language}/news/tags/patch-notes/`;
-        this.patch_note = `https://playvalorant.com/${lowerCase_language}/news/game-updates`;
+        this.baseUrl = {
+            default: `https://playvalorant.com/${lowerCase_language}`,
+            data: `https://playvalorant.com/page-data/${lowerCase_language}`
+        };
     }
 
-    /**
-     *
-     * @param {string} language Language
-     * @returns {string} Url of Json data
-     */
-    public static getJsonUrl(language?: Locale.Identify): string {
-        const _myPatchNote = new PatchNote(language);
-
-        return `${_myPatchNote.json}`;
-    }
-
-    /**
-     *
-     * @param {string} patch Version
-     * @param {string} language Language
-     * @returns {string} Url of Patch note
-     */
-    public static getUrl(patch?: PatchNote.Version, language?: Locale.Identify): string {
-        const _myPatchNote = new PatchNote(language);
-
-        if (!patch) {
-            return `${_myPatchNote.patch_list}`;
-        }
-
+    private static parse(patch: ValPatchNote.Version, replaceWith = "."): string {
         const split_patch: Array<string> = patch.split(".");
         if (split_patch.length !== 2) {
             throw new ValError({
@@ -62,12 +40,54 @@ export class PatchNote {
                 message: "Invalid patch number",
                 data: patch
             });
-        } else if (split_patch.at(1) !== "0" && String(split_patch.at(1)).length === 1) {
-            split_patch[1] = `0${split_patch[1]}`;
-
-            patch = `${Number(split_patch.at(0))}.${String(split_patch.at(1))}` as PatchNote.Version;
         }
 
-        return `${_myPatchNote.patch_note}/valorant-patch-notes-${patch.replace(".", "-")}`;
+        // * first Value
+
+        split_patch[0] = Number(split_patch[0]).toString();
+
+        // * second value
+
+        if (split_patch[1] === "00") {
+            // 5.00 --> 5.0
+
+            split_patch[1] = "0";
+        } else if (split_patch[1] !== "0") {
+            // 5.1 --> 5.01
+
+            if (String(split_patch[1]).length === 1) {
+                split_patch[1] = `0${split_patch[1]}`;
+            }
+        }
+
+        return split_patch.join(replaceWith);
+    }
+
+    // static
+
+    /**
+     *
+     * @param {string} patch Version
+     * @param {string} language Language
+     * @returns {string} Url
+     */
+    public static getUrl(patch?: ValPatchNote.Version, language?: Locale.Identify): string {
+        const _patchNote = new ValPatchNote(language);
+
+        if (!patch) {
+            return `${_patchNote.baseUrl.default}/news/tags/patch-notes/`;
+        }
+
+        return `${_patchNote.baseUrl.default}/news/game-updates/valorant-patch-notes-${ValPatchNote.parse(patch, "-")}`;
+    }
+
+    /**
+     *
+     * @param {string} language Language
+     * @returns {string} Url
+     */
+    public static getContentUrl(language?: Locale.Identify): string {
+        const _patchNote = new ValPatchNote(language);
+        return `${_patchNote.baseUrl.data}/news/tags/patch-notes/page-data.json`;
     }
 }